+++ app/actions/auth.ts
@@ -1,0 +1,11 @@
+'use server';
+
+import { signIn, signOut } from '@/app/api/auth/[...nextauth]/route';
+
+export async function keycloakSignIn(callbackUrl: string) {
+  await signIn('keycloak', { callbackUrl: '/create-ingest' });
+}
+
+export async function keycloakSignOut(callbackUrl: string) {
+  await signOut({ redirect: false });
+}
