+++ app/api/auth/[...nextauth]/route.ts
@@ -1,20 +1,27 @@
-import NextAuth, { NextAuthConfig } from 'next-auth';
+import NextAuth from 'next-auth';
 import KeycloakProvider from 'next-auth/providers/keycloak';
 
-const providers =
-  process.env.NEXT_PUBLIC_DISABLE_AUTH === 'true'
-    ? []
-    : [
-        KeycloakProvider({
-          clientId: process.env.KEYCLOAK_CLIENT_ID!,
-          clientSecret: process.env.KEYCLOAK_CLIENT_SECRET!,
-          issuer: process.env.KEYCLOAK_ISSUER!,
-        }),
-      ];
+const providers = [];
 
-const authOptions = {
-  providers,
+if (process.env.NEXT_PUBLIC_DISABLE_AUTH !== 'true') {
+  providers.push(
+    KeycloakProvider({
+      clientId: process.env.KEYCLOAK_CLIENT_ID!,
+      clientSecret: process.env.KEYCLOAK_CLIENT_SECRET,
+      issuer: process.env.NEXT_PUBLIC_KEYCLOAK_ISSUER!,
+    })
+  );
+}
+
+export const {
+  handlers: { GET, POST },
+  auth,
+  signIn,
+  signOut,
+} = NextAuth({
   secret: process.env.NEXTAUTH_SECRET,
-} satisfies Parameters<typeof NextAuth>[0];
-
-const handler = NextAuth(authOptions);
+  providers,
+  session: {
+    strategy: 'jwt',
+  },
+});
