+++ app/upload/_components/UploadClient.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import AppLayout from '@/components/Layout';
+import ThumbnailUploader from '@/components/ThumbnailUploader';
+
+function UploadClient() {
+  return (
+    <AppLayout>
+      <ThumbnailUploader />
+    </AppLayout>
+  );
+}
+
+export default UploadClient;
