+++ components/LogoutButton.tsx
@@ -3,7 +3,9 @@
 import React from 'react';
 import { Button, Tooltip } from 'antd';
 import { LogoutOutlined } from '@ant-design/icons';
-import { signOut } from 'next-auth/react';
+import { useSearchParams, useRouter } from 'next/navigation';
+import { useTransition } from 'react';
+import { keycloakSignOut } from '@/app/actions/auth';
 
 const collapsedStyle = {
   marginTop: '16px',
@@ -19,13 +21,30 @@ const expandedStyle = {
   justifyContent: 'flex-start',
 };
 
-const LogoutButton = ({ collapsed }: { collapsed: boolean }) => {
+interface LogoutButtonProps {
+  collapsed: boolean;
+}
+
+const LogoutButton = ({ collapsed }: LogoutButtonProps) => {
+  const searchParams = useSearchParams();
+  const router = useRouter();
+  const [isPending, startTransition] = useTransition();
+  const callbackUrl = searchParams.get('callbackUrl') || '/login';
+
+  const handleLogout = () => {
+    startTransition(async () => {
+      await keycloakSignOut(callbackUrl);
+      const keycloakLogoutUrl = `${process.env.NEXT_PUBLIC_KEYCLOAK_ISSUER}/protocol/openid-connect/logout?redirect_uri=${encodeURIComponent(window.location.origin + callbackUrl)}`;
+      window.location.href = keycloakLogoutUrl;
+    });
+  };
+
   return (
     <Tooltip placement="right" title={collapsed ? 'Sign out' : ''}>
       <Button
-        variant="solid"
-        color="danger"
-        onClick={() => signOut({ callbackUrl: '/' })}
+        type="primary" // Changed from variant="solid" and color="danger" for Ant Design Button
+        danger
+        onClick={handleLogout}
         icon={<LogoutOutlined />}
         block
         style={collapsed ? collapsedStyle : expandedStyle}
